Overhauling a lot of details to simplify
