+++ docker/noetic_compose.yaml
@@ -2,7 +2,7 @@ services:
   rosbridge_noetic:
     image: carter12s/roslibrust-ci-noetic:rust_1_90
     # network_mode host required for ros1 testing
-    # network_mode: host
+    network_mode: host
     # ports:
     # - "9090:9090"
     # Pass through the ros master port for native ros1 testing
